Outdated dependency 'vis' Fixes #807
