+++ front/src/routes/integration/all/zwave/network-page/actions.js
@@ -1,48 +1,7 @@
-import vis from 'vis';
 import get from 'get-value';
 
 import { RequestStatus } from '../../../../../utils/consts';
 import { ERROR_MESSAGES } from '../../../../../../../server/utils/constants';
-
-function renderGraph(zwaveNodes) {
-  const nodeData = zwaveNodes.map(node => ({
-    id: node.id,
-    label: node.product
-  }));
-  let nodes = new vis.DataSet(nodeData);
-
-  const edgeData = [];
-
-  const alreadyIn = {};
-
-  zwaveNodes.forEach(node => {
-    node.neighbors.forEach(neighborId => {
-      if (!alreadyIn[`${neighborId}-${node.id}`]) {
-        edgeData.push({
-          from: node.id,
-          to: neighborId
-        });
-        alreadyIn[`${node.id}-${neighborId}`] = true;
-      }
-    });
-  });
-
-  // create an array with edges
-  let edges = new vis.DataSet(edgeData);
-
-  // create a network
-  let container = document.getElementById('zwave-network-graph');
-
-  // provide the data in the vis format
-  let data = {
-    nodes,
-    edges
-  };
-  let options = {};
-
-  // initialize your network!
-  return new vis.Network(container, data, options);
-}
 
 const actions = store => ({
   async getNeighbors(state) {
@@ -55,7 +14,6 @@ const actions = store => ({
         zwaveNodesNeighbors,
         zwaveGetNeighborsStatus: RequestStatus.Success
       });
-      renderGraph(zwaveNodesNeighbors);
     } catch (e) {
       const responseMessage = get(e, 'response.data.message');
       if (responseMessage === ERROR_MESSAGES.SERVICE_NOT_CONFIGURED) {
